services/scraper/requirements.txt
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
celery[redis]>=5.3.0,<6.0
redis>=5.0.0,<6.0
requests>=2.31.0,<3.0
beautifulsoup4>=4.12.0,<5.0
lxml>=5.0.0,<6.0
psycopg2-binary>=2.9.0,<3.0
sqlalchemy>=2.0.0,<3.0
python-dateutil>=2.8.0,<3.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0
httpx>=0.26.0,<1.0
python-dotenv>=1.0.0,<2.0
robotexclusionrulesparser>=1.7.0,<2.0
eval-type-backport>=0.2.0,<1.0
